fix: rename things to api- since they are for api traces
