+++ projects/distributed-tracing/src/shared/dashboard/widgets/trace-detail/trace-detail-widget.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { LoadAsyncModule, SummaryValueModule, TitledContentModule } from '@hypertrace/components';
 import { DashboardCoreModule } from '@hypertrace/hyperdash-angular';
 import { SpanDetailModule } from '../../../components/span-detail/span-detail.module';
-import { TraceDetailDataSourceModel } from './data/trace-detail-data-source.model';
+import { ApiTraceDetailDataSourceModel } from './data/api-trace-detail-data-source.model';
 import { TraceDetailWidgetRendererComponent } from './trace-detail-widget-renderer.component';
 import { TraceDetailWidgetModel } from './trace-detail-widget.model';
 
@@ -11,7 +11,7 @@ import { TraceDetailWidgetModel } from './trace-detail-widget.model';
   declarations: [TraceDetailWidgetRendererComponent],
   imports: [
     DashboardCoreModule.with({
-      models: [TraceDetailWidgetModel, TraceDetailDataSourceModel],
+      models: [TraceDetailWidgetModel, ApiTraceDetailDataSourceModel],
       renderers: [TraceDetailWidgetRendererComponent]
     }),
     CommonModule,
