+++ projects/observability/src/pages/apis/api-detail/traces/api-trace-list.dashboard.ts
@@ -67,7 +67,7 @@ export const apiTraceListDashboard = {
     'child-template': {
       type: 'trace-detail-widget',
       data: {
-        type: 'trace-detail-data-source',
+        type: 'api-trace-detail-data-source',
         // tslint:disable-next-line: no-invalid-template-strings
         trace: '${row}',
         attributes: ['requestUrl']
