+++ projects/observability/src/pages/apis/backend-detail/traces/backend-trace-list.dashboard.ts
@@ -56,7 +56,7 @@ export const backendTraceListDashboard = {
     'child-template': {
       type: 'trace-detail-widget',
       data: {
-        type: 'trace-detail-data-source',
+        type: 'api-trace-detail-data-source',
         // tslint:disable-next-line: no-invalid-template-strings
         trace: '${row}'
       }
