+++ projects/observability/src/pages/explorer/explorer-dashboard-builder.test.ts
@@ -116,7 +116,7 @@ describe('Explorer dashboard builder', () => {
             'child-template': {
               type: 'trace-detail-widget',
               data: {
-                type: 'trace-detail-data-source',
+                type: 'api-trace-detail-data-source',
                 // tslint:disable-next-line: no-invalid-template-strings
                 trace: '${row}',
                 attributes: ['requestUrl']
