+++ projects/observability/src/pages/explorer/explorer-dashboard-builder.ts
@@ -131,7 +131,7 @@ export class ExplorerDashboardBuilder {
       'child-template': {
         type: 'trace-detail-widget',
         data: {
-          type: 'trace-detail-data-source',
+          type: 'api-trace-detail-data-source',
           // tslint:disable-next-line: no-invalid-template-strings
           trace: '${row}',
           attributes: ['requestUrl']
